==> codes/1.0_Data_Collection.ipynb <==
# 1. Data Collection Trial and Error
*Author: Boom*

### City Data Frame

import pandas as pd

# Read .csv
cities = pd.read_csv('./US_City_Data/USCFinal.csv', index_col=0)

# Filter out only NE cities
NE_cities = cities[cities['state_id'].isin(["ME", "VT", "NH","MA", "CT", "RI"])]

# Extract only 
NE_cities = NE_cities.loc[:,['city', 'state_id', 'state_name', 'county_name', 'lat', 'lng']]

# View
NE_cities.head()

# Save down
NE_cities.to_csv("./datasets/new_england_cities_geo-data.csv")

### TwitterScraper

Spoke to Tucker and he said if there's no way around this, we can turn this into a "spam filter"-like project where we can idenfity posts that are actually about power outages.

Alternatively, I suggested we could add a column to our data frame where we make up a location, which he says is fine for proof-of-concept. He said that's even better since it'll give us a chance to work with geospatial data visualization.

# Install Twitter Scraper
!pip install twitterscraper
import twitterscraper

# Load Packages
from twitterscraper import query_tweets
import datetime

# Create empty template tweet dataframe to later populate 
tweet_df =pd.DataFrame(columns=["id","text","timestamp"])

# Check
tweet_df

# Query Tweets between specified dates
list_of_tweets = query_tweets("blackout OR blackouts OR outage OR outages OR power outage",
                              begindate = datetime.date(2018,1,1),
                              enddate = datetime.date(2018,12,31),
                              poolsize = 1)

# Extract features of tweets to populate dataframe:
for row, tweet in enumerate(list_of_tweets):
    tweet_df.loc[row,'id'] = tweet.id
    tweet_df.loc[row,'text'] = tweet.text
    tweet_df.loc[row,'timestamp'] = tweet.timestamp

# Examine
print(tweet_df.shape)

# Save as .csv
# tweet_df.to_csv("./dirty_tweets_20180101-20181231.csv")
