==> codes/2.0_Data_Cleaning.ipynb <==
# 2. Data Cleaning
*Author: Boom*

### Import Packages and Data

# Import Packages
import numpy as np
import pandas as pd
import scipy.stats as stats

# Read in pulled data
city_info = pd.read_csv("./datasets/new_england_cities_geo-data.csv")
tweets_df = pd.read_csv("./datasets/dirty_tweets_20180101-20181231.csv")

# Clean these data frames by getting rid of unnamed column
city_info.drop(columns=["Unnamed: 0"], inplace=True)
tweets_df.drop(columns=["Unnamed: 0"], inplace=True)

### Simulate City Location for each Tweet
- At this stage, the purpose of this project is to provide proof-of-concept.
- Due to our numerous attempts to determine the optimal way to pull relevant Tweets with location, Twitter's API (via `twitterscrape`) has locked us out thereby preventing us from pulling any more Tweets.
- These tweets we currently have here are relevant to power outages, but do not have informative location information.
- Therefore, in order to proceed with modeling and geospatial visualization, we need to simulate the locations by randomly sampling from the unique list of cities.

# Get unique list of cities
city_list = list(city_info['city'].unique())

# Randomly assign a city to each tweet
np.random.seed(42)
tweets_df["city"] = np.random.choice(city_list, size=tweets_df.shape[0], replace=True)

# Merge rest of city_info dataframe to tweets_df
tweets_location_df = pd.merge(tweets_df, city_info, how="left", on="city")

# Get rid of duplicate Tweets
tweets_location_df.drop_duplicates(subset='text', keep="last", inplace=True)

# Set Tweet ID and index and get rid of ID column
tweets_location_df.set_index(tweets_location_df['id'], inplace=True)
tweets_location_df.drop(columns=["id"], inplace=True)

# Peak
tweets_location_df.head(5)

print("We now have", tweets_location_df.shape[0],"unique Tweets remaining")

# Save as .csv
tweets_location_df.to_csv("./datasets/clean_tweets_20180101-20181231.csv")
