==> codes/3.1_Keywords.ipynb <==
# Proposed method for extending power outage keywords:
Henry

## Basics:
The functions below will use Word2Vec to find terminology related to a list of words I've generated, which I believe will help identify tweets from areas affected by power outages.  

The list of words I'm searching through came from an article about power-outage terminology - terms that affected people might become familiar with or use often - and supplemented by common sense.  

The approach here is simple: we want the terms I've identified, plus any words that are considered 40% or more related to them.  This will populate a list that we can use to search tweets. 

# Libraries:

import pandas as pd
import numpy as np
import time
import gensim

## Note: The file below is too large to transmit over Git, but you should both have it locally in your word vectors repo!  Grab it and run to see what I worked up here.

# Wiki text-data: takes ~ 3 minutes to load:

t0 = time.time()
model = gensim.models.KeyedVectors.load_word2vec_format('./lexvec.enwiki+newscrawl.300d.W.pos.vectors.gz')
print(time.time() - t0)

# My identified terms:

words = ['blackout','power','outage','electric','electricity','electrical','transformer','watt','wattage','arc',
         'circuit','breaker','cable','fault','conductor','fuse','riser','insulator','meter',
         'interruption','maintenance','relay','grid','severe','weather','storm','substation','surge','switch',
         'switchyard','station','transmission','system','lines','line','frequency','voltage']

## Function 1: Dictionary layout:
- Use to edit down terms
- Some terms are clearly returning more technical electrical terms than we would want

banker = {}
for word in words:
    arr = pd.DataFrame(model.most_similar(word,topn=20),columns=['word','val'])
    arr = arr[arr['val']>.39]
    banker[word] = list(arr.word)
    
banker

# Function 2: List Layout:
bank = []
for word in words:
    arr = pd.DataFrame(model.most_similar(word,topn=20),columns=['word','val'])
    arr = arr[arr['val']>.39]
    for i in arr.word:
        bank.append(i)

# Add origional terms to list
for word in words:
    bank.append(word)

#print(bank)

## Conclusions:
**Based on the output of function 1:**

I propose we eliminate some of the more technical terms, like "electric",'watt','arc'... and then see if the terms generated are more managable.  Let me know what you think tomorrow, I'm going to bed!
